backend/pyproject.toml
[project]
name = "qa-assignment-backend"
version = "1.0.0"
description = "FastAPI backend for QA Assignment"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.23",
    "alembic==1.13.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
